+++ kafka/checkpoint.go
@@ -1,50 +1,66 @@
 package kafka
 
 import (
+	"fmt"
 	"strconv"
 	"time"
 
 	"github.com/Shopify/sarama"
+	"github.com/araddon/dateparse"
 
 	"github.com/xitonix/trubka/internal"
 )
 
-// CheckpointPair represents a pair of From/To checkpoints.
-type CheckpointPair struct {
-	// From Specifies where to start consuming from.
-	From *Checkpoint
-	// To Specifies when the consumer must stop.
-	To *Checkpoint
+type checkpointMode int8
+
+const (
+	// Predefined Sarama values (eg. newest, oldest etc)
+	predefinedMode checkpointMode = iota
+	explicitMode
+	localMode
+	timestampMode
+)
+
+type checkpointPair struct {
+	from *checkpoint
+	to   *checkpoint
 }
 
-// Checkpoint represents a point in time or offset, from which the consumer has to start consuming from the specified topic.
-type Checkpoint struct {
-	offset      int64
-	at          time.Time
-	isTimeBased bool
+type checkpoint struct {
+	offset int64
+	at     time.Time
+	mode   checkpointMode
 }
 
-func newCheckpoint(rewind bool) *Checkpoint {
+func newPredefinedCheckpoint(rewind bool) *checkpoint {
 	offset := sarama.OffsetNewest
 	if rewind {
 		offset = sarama.OffsetOldest
 	}
-	return &Checkpoint{
+	return &checkpoint{
 		offset: offset,
+		mode:   predefinedMode,
 	}
 }
 
-func newOffsetCheckpoint(offset int64) *Checkpoint {
+func newLocalCheckpoint() *checkpoint {
+	return &checkpoint{
+		mode: localMode,
+	}
+}
+
+func newExplicitCheckpoint(offset int64) *checkpoint {
 	if offset < sarama.OffsetOldest {
 		offset = sarama.OffsetOldest
 	}
-	return &Checkpoint{
+	return &checkpoint{
 		offset: offset,
+		mode:   explicitMode,
 	}
 }
 
 // newTimeCheckpoint creates a new checkpoint and sets the offset to the milliseconds of the given time and the mode to MillisecondsOffsetMode.
-func newTimeCheckpoint(at time.Time) *Checkpoint {
+func newTimeCheckpoint(at time.Time) *checkpoint {
 	var offsetMilliSeconds int64
 	switch {
 	case at.IsZero():
@@ -52,18 +68,16 @@ func newTimeCheckpoint(at time.Time) *Checkpoint {
 	default:
 		offsetMilliSeconds = at.UnixNano() / 1000000
 	}
-	return &Checkpoint{
-		isTimeBased: true,
-		offset:      offsetMilliSeconds,
-		at:          at,
+	return &checkpoint{
+		mode:   timestampMode,
+		offset: offsetMilliSeconds,
+		at:     at,
 	}
 }
 
-// OffsetString returns the string representation of the time offset in `02-01-2006T15:04:05.999999999` format if in
-// MillisecondsOffsetMode mode, otherwise returns the string representation of the offset value.
-func (c *Checkpoint) OffsetString() string {
-	if c.isTimeBased {
-		return internal.FormatTimeForHuman(c.at)
+func (c *checkpoint) String() string {
+	if c.mode == timestampMode {
+		return internal.FormatTime(c.at)
 	}
 	switch c.offset {
 	case sarama.OffsetNewest:
@@ -75,17 +89,56 @@ func (c *Checkpoint) OffsetString() string {
 	}
 }
 
-func (c *Checkpoint) after(cp *Checkpoint) bool {
+func (c *checkpoint) after(cp *checkpoint) bool {
 	if c == nil || cp == nil {
 		return false
 	}
 
-	if c.isTimeBased {
-		if cp.isTimeBased {
+	if c.mode == timestampMode {
+		if cp.mode == timestampMode {
 			return c.at.After(cp.at)
 		}
 		return false
 	}
 
 	return c.offset > cp.offset
+}
+
+func parseCheckpoint(value string, isStopOffset bool) (*checkpoint, error) {
+	if len(value) == 0 {
+		return nil, fmt.Errorf("start/stop checkpoint value cannot be empty")
+	}
+	t, err := dateparse.ParseAny(value)
+	if err == nil {
+		return newTimeCheckpoint(t), nil
+	}
+
+	switch value {
+	case "local", "stored":
+		if isStopOffset {
+			return nil, invalidToValue(value)
+		}
+		return newLocalCheckpoint(), nil
+	case "newest", "latest", "end":
+		if isStopOffset {
+			return nil, invalidToValue(value)
+		}
+		return newPredefinedCheckpoint(false), nil
+	case "oldest", "earliest", "beginning", "start":
+		if isStopOffset {
+			return nil, invalidToValue(value)
+		}
+		return newPredefinedCheckpoint(true), nil
+	}
+
+	offset, err := parseInt(value, "offset")
+	if err != nil {
+		return nil, err
+	}
+
+	return newExplicitCheckpoint(offset), nil
+}
+
+func invalidToValue(value string) error {
+	return fmt.Errorf("'%s' is not an acceptable stop condition", value)
 }