+++ kafka/offset.go
@@ -13,6 +13,7 @@ type Offset struct {
 	Latest int64
 	// Current the current value of the local or consumer group offset. This is where the consumer up to.
 	Current int64
+	StopAt  *checkpoint `json:"-"`
 }
 
 // Lag calculates the lag between the latest and the current offset values.